+++ apps/course-platform-analog/src/app/pages/login.page.ts
@@ -1,6 +1,6 @@
 import { RouteMeta } from '@analogjs/router';
 import { LoginComponent } from '@course-platform/course-client/feature';
-import { redirectIfLoggedInServerGuard } from '@course-platform/course-client/shared/domain';
+import { redirectIfLoggedInServerGuard } from '@course-platform/shared/auth/domain';
 
 export const routeMeta: RouteMeta = {
   title: 'Login',
