+++ libs/course-admin/settings/feature/src/lib/tabs/profile/profile.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { Observable } from 'rxjs';
+import { Observable, from } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 import { User } from '@angular/fire/auth';
 import { SharedModule } from '@course-platform/course-admin/shared/ui';
-import { ProfileService } from '@course-platform/course-client/shared/domain';
-import { AuthService } from '@course-platform/shared/auth/domain';
+import { AuthFBService } from '@course-platform/shared/auth/domain';
+import { ProfileFBService } from '@course-platform/course-client/shared/domain';
 
 @Component({
   selector: 'app-profile',
@@ -21,13 +21,13 @@ export class ProfileComponent implements OnInit {
   changePwForm!: FormGroup;
   errorMessage!: string;
   constructor(
-    private profileService: ProfileService,
+    private profileService: ProfileFBService,
     private formBuilder: FormBuilder,
-    private authService: AuthService
+    private authService: AuthFBService
   ) {}
 
   ngOnInit() {
-    this.user$ = this.profileService.getUserProfile();
+    this.user$ = from(this.profileService.getUserProfile());
 
     this.profileForm$ = this.user$.pipe(
       map((user) =>
@@ -71,6 +71,6 @@ export class ProfileComponent implements OnInit {
   }
 
   onLogout() {
-    this.authService.doLogout();
+    this.authService.signOut();
   }
 }
