+++ libs/course-admin/shared/domain/src/index.ts
@@ -1,4 +1,3 @@
-export * from './lib/auth/redirect-if-authenticated.service';
 export * from './lib/course-admin-facade.service';
 export * from './lib/course-admin-shared-domain.module';
 export * from './lib/environments/environment';
