+++ libs/course-admin/shared/domain/src/lib/course-admin-shared-domain.module.ts
@@ -1,9 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule, Optional, SkipSelf } from '@angular/core';
-import { SharedDataAccessModule } from '@course-platform/shared/domain';
 
 @NgModule({
-  imports: [CommonModule, SharedDataAccessModule],
+  imports: [CommonModule],
 })
 export class CourseAdminSharedDomainModule {
   constructor(
