+++ libs/course-admin/shell/src/lib/app.routing.ts
@@ -1,8 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { RedirectIfAuthenticatedResolver } from '@course-platform/course-admin/shared/domain';
-import { AuthFBGuard } from '@course-platform/shared/auth/domain';
+import {
+  AuthFBGuard,
+  RedirectIfLoggedInResolver,
+} from '@course-platform/shared/auth/domain';
 import { SchoolIdResolver } from '@course-platform/shared/domain';
 import { LayoutComponent } from './layout/layout.component';
 
@@ -18,7 +20,7 @@ const routes: Routes = [
     children: [
       {
         path: 'login',
-        resolve: [RedirectIfAuthenticatedResolver],
+        resolve: [RedirectIfLoggedInResolver],
         loadComponent: () =>
           import('@course-platform/course-admin/login/feature').then(
             (m) => m.LoginComponent
