--- libs/course-client/shared/domain/src/auth/index.ts
@@ -1,4 +1,0 @@
-// start:ng42.barrel
-export * from './redirect-if-logged-in.service';
-export * from './redirect-if-logged-out.service';
-// end:ng42.barrel
