--- libs/course-client/shared/domain/src/auth/redirect-if-logged-out.service.ts
@@ -1,36 +1,0 @@
-import { Injectable, PLATFORM_ID, inject } from '@angular/core';
-import { Auth } from '@angular/fire/auth';
-import { ActivatedRouteSnapshot, Resolve, Router } from '@angular/router';
-import { Observable, of } from 'rxjs';
-import { first, map } from 'rxjs/operators';
-
-import { isPlatformServer } from '@angular/common';
-import { UserService } from '@course-platform/shared/auth/domain';
-import { SsrCookieService } from 'ngx-cookie-service-ssr';
-
-@Injectable({ providedIn: 'root' })
-export class RedirectIfLoggedOutResolver implements Resolve<boolean> {
-  private platformId = inject(PLATFORM_ID);
-  constructor(
-    public afAuth: Auth,
-    public userService: UserService,
-    private router: Router
-  ) {}
-  resolve(route: ActivatedRouteSnapshot): Observable<boolean> {
-    if (isPlatformServer(this.platformId)) {
-      return of(true);
-    }
-
-    return this.userService.getCurrentUser().pipe(
-      first(),
-      map((currentUser) => {
-        if (!currentUser) {
-          console.log('Not authenticated, redirecting to login on client.');
-          this.router.navigate(['login']);
-          return false;
-        }
-        return true;
-      })
-    );
-  }
-}
