+++ libs/course-client/shared/domain/src/core.module.ts
@@ -1,10 +1,8 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 
-import { SharedDataAccessModule } from '@course-platform/shared/domain';
-
 @NgModule({
-  imports: [SharedDataAccessModule],
+  imports: [],
 })
 export class CoreModule {
   constructor(
