+++ libs/course-client/shared/domain/src/index.ts
@@ -1,9 +1,4 @@
 // start:ng42.barrel
-export * from './auth';
-export * from './auth/redirect-if-logged-in-server.guard';
-export * from './auth/redirect-if-logged-in.service';
-export * from './auth/auth.guard';
-export * from './auth/redirect-if-logged-out.service';
 export * from './core.module';
 export * from './course';
 export * from './course-client-domain.module';
