+++ libs/course-client/shared/domain/src/profile/index.ts
@@ -1,3 +1,3 @@
 // start:ng42.barrel
-export * from './profile.service';
+export * from './profile-fb.service';
 // end:ng42.barrel
