+++ libs/course-client/shared/ui/src/layout/topbar/topbar.component.html
@@ -4,6 +4,7 @@
     <a class="branding spacer" [href]="homeUrl$ | async">
       <img src="assets/logo.png" alt="logo" class="inline" />
       {{ 'topbar.title' | translate }}
+      dfd
     </a>
 
     <ul class="menu-items" *ngIf="loggedIn$ | async">
