+++ libs/course-client/shared/ui/src/layout/topbar/topbar.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Observable, from } from 'rxjs';
+import { map, tap } from 'rxjs/operators';
 
 import { CourseClientFacade } from '@course-platform/course-client/shared/domain';
-import { UserService } from '@course-platform/shared/auth/domain';
+import {
+  AuthFBService,
+  AuthService,
+} from '@course-platform/shared/auth/domain';
 
 interface NavigationItem {
   link: string;
@@ -17,19 +20,17 @@ interface NavigationItem {
   templateUrl: './topbar.component.html',
   styleUrls: ['./topbar.component.scss'],
 })
-export class TopbarComponent implements OnInit {
+export class TopbarComponent {
   languages = ['en'];
   homeUrl$!: Observable<string>;
   navigationItems!: NavigationItem[];
   loggedIn$!: Observable<boolean>;
 
   constructor(
-    private userService: UserService,
+    private authService: AuthService,
     private courseClientFacade: CourseClientFacade
-  ) {}
-
-  ngOnInit() {
-    this.loggedIn$ = this.userService.isLoggedIn$;
+  ) {
+    this.loggedIn$ = this.authService.isLoggedIn();
     this.homeUrl$ = this.courseClientFacade.courseId$.pipe(
       map((courseId) => `courses/${courseId ? courseId : ''}`)
     );
