+++ libs/shared/auth/domain/src/index.ts
@@ -1,8 +1,9 @@
-export * from './lib/auth/auth-server.interceptor';
-export * from './lib/auth/auth-fb.guard';
-export * from './lib/auth/auth.interceptor';
-export * from './lib/auth/auth.service';
-export * from './lib/auth/user-server.service';
-export * from './lib/auth/user.service';
-export * from './lib/auth/auth-sb.service';
-export * from './lib/auth/auth-client';
+export * from './lib/auth/interceptors/auth-fb-server.interceptor';
+export * from './lib/auth/interceptors/auth-fb.interceptor';
+export * from './lib/auth/guards/auth-fb.guard';
+export * from './lib/auth/services/auth-fb.service';
+export * from './lib/auth/services/auth.service';
+export * from './lib/auth/services/auth-sb.service';
+export * from './lib/auth/guards/redirect-if-logged-in-server.guard';
+export * from './lib/auth/guards/redirect-if-logged-in.service';
+export * from './lib/auth/guards/redirect-if-logged-out.service';
