--- libs/shared/auth/domain/src/lib/auth/auth-client.ts
@@ -1,14 +1,0 @@
-import { AuthClient } from '@supabase/auth-js';
-/// <reference types="vite/client" />
-
-const AUTH_URL = `${import.meta.env['VITE_SUPABASE_URL']!}/auth/v1`;
-const AUTH_HEADERS = {
-  Authorization: `Bearer ${import.meta.env['VITE_SUPABASE_KEY']!}`,
-  apikey: `${import.meta.env['VITE_SUPABASE_KEY']!}`,
-};
-
-export const authClient = new AuthClient({
-  headers: AUTH_HEADERS,
-  url: AUTH_URL,
-  fetch: fetch,
-});
