+++ libs/shared/auth/domain/src/lib/auth/guards/auth-fb.guard.ts
@@ -1,18 +1,17 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable, of } from 'rxjs';
+import { Observable, from, of } from 'rxjs';
 import { catchError, first, map } from 'rxjs/operators';
-
-import { UserService } from './user.service';
+import { AuthFBService } from '../services/auth-fb.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthFBGuard {
-  constructor(public userService: UserService, private router: Router) {}
+  constructor(public authFBService: AuthFBService, private router: Router) {}
 
   canActivate(): Observable<boolean> {
-    return this.userService.currentUser$.pipe(
+    return from(this.authFBService.getFBUser()).pipe(
       first(),
       map((currentUser) => {
         if (!currentUser) {
