+++ libs/shared/domain/src/index.ts
@@ -1,5 +1,5 @@
 export * from './resources';
-export * from './shared-data-access.module';
 export * from './ngrx/router';
 export * from './school-id';
+export * from './firebase.module';
 export * from './custom-domain';
