+++ libs/shared/domain/src/resources/index.ts
@@ -1,4 +1,5 @@
 export * from './course-resources.service';
 export * from './course-resources-trpc.service';
+export * from './course-resources-fb.service';
 export * from './endpoints';
 export * from './graphql';
