+++ libs/shared/domain/trpc-server/src/lib/trpc/routers/user-service.ts
@@ -1,4 +1,3 @@
-import { authClient } from '../context';
 import { supbaseAdmin } from '../supabase-admin';
 
 export interface CreateUserResponseDTO {
